Revamp maven project + webapp module
