+++ pom.xml
@@ -6,18 +6,57 @@
 
     <groupId>syncworkout</groupId>
     <artifactId>syncworkout-parent</artifactId>
+    <packaging>pom</packaging>
+    <name>SyncWorkout - Parent</name>
     <version>1.0-SNAPSHOT</version>
 
+    <modules>
+        <module>webapp</module>
+    </modules>
+
     <properties>
+        <version.appengine.app>1</version.appengine.app>
+        <version.appengine.target>1.7.5</version.appengine.target>
         <version.jersey>1.18.3</version.jersey>
+        <version.jstl>1.2</version.jstl>
+        <version.servlet-api>2.5</version.servlet-api>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     </properties>
 
-    <dependencies>
-        <dependency>
-            <groupId>com.sun.jersey</groupId>
-            <artifactId>jersey-server</artifactId>
-            <version>${version.jersey}</version>
-        </dependency>
-    </dependencies>
+    <dependencyManagement>
+        <dependencies>
+            <dependency>
+                <groupId>com.google.appengine</groupId>
+                <artifactId>appengine-api-1.0-sdk</artifactId>
+                <version>${version.appengine.target}</version>
+            </dependency>
+            <dependency>
+                <groupId>javax.servlet</groupId>
+                <artifactId>servlet-api</artifactId>
+                <version>${version.servlet-api}</version>
+                <scope>provided</scope>
+            </dependency>
+            <dependency>
+                <groupId>jstl</groupId>
+                <artifactId>jstl</artifactId>
+                <version>${version.jstl}</version>
+            </dependency>
+            <dependency>
+                <groupId>com.sun.jersey</groupId>
+                <artifactId>jersey-server</artifactId>
+                <version>${version.jersey}</version>
+            </dependency>
+            <dependency>
+                <groupId>com.google.appengine</groupId>
+                <artifactId>appengine-testing</artifactId>
+                <version>${version.appengine.target}</version>
+            </dependency>
+            <dependency>
+                <groupId>com.google.appengine</groupId>
+                <artifactId>appengine-api-stubs</artifactId>
+                <version>${version.appengine.target}</version>
+            </dependency>
+        </dependencies>
+    </dependencyManagement>
 
 </project>
