--- src/main/java/example/HelloWorldClient.java
@@ -1,8 +1,0 @@
-package example;/**
- * Created by Jason on 15/02/2015.
- */
-public class HelloWorldClient {
-  public static void main(String[] argv) {
-
-  }
-}
