+++ webapp/nbactions.xml
@@ -1,0 +1,52 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<actions>
+    <action>
+        <actionName>CUSTOM-appengine:devserver</actionName>
+        <displayName>appengine:devserver</displayName>
+        <goals>
+            <goal>appengine:devserver</goal>
+        </goals>
+    </action>
+    <action>
+        <actionName>CUSTOM-appengine:update</actionName>
+        <displayName>appengine:update</displayName>
+        <goals>
+            <goal>appengine:update</goal>
+        </goals>
+    </action>
+    <action>
+        <actionName>CUSTOM-appengine:rollback</actionName>
+        <displayName>appengine:rollback</displayName>
+        <goals>
+            <goal>appengine:rollback</goal>
+        </goals>
+    </action>
+    <action>
+        <actionName>CUSTOM-appengine:update_cron</actionName>
+        <displayName>appengine:update_cron</displayName>
+        <goals>
+            <goal>appengine:update_cron</goal>
+        </goals>
+    </action>
+    <action>
+        <actionName>CUSTOM-appengine:update_dos</actionName>
+        <displayName>appengine:update_dos</displayName>
+        <goals>
+            <goal>appengine:update_dos</goal>
+        </goals>
+    </action>
+    <action>
+        <actionName>CUSTOM-appengine:update_indexes</actionName>
+        <displayName>appengine:update_indexes</displayName>
+        <goals>
+            <goal>appengine:update_indexes</goal>
+        </goals>
+    </action>
+    <action>
+        <actionName>CUSTOM-appengine:update_queues</actionName>
+        <displayName>appengine:update_queues</displayName>
+        <goals>
+            <goal>appengine:update_queues</goal>
+        </goals>
+    </action>
+</actions>
