+++ webapp/src/main/webapp/WEB-INF/appengine-web.xml
@@ -1,12 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <appengine-web-app xmlns="http://appengine.google.com/ns/1.0">
-    <application>syncworkouts</application>
+    <application>syncworkout-webapp</application>
     <version>${appengine.app.version}</version>
     <threadsafe>true</threadsafe>
-
-
 
     <system-properties>
         <property name="java.util.logging.config.file" value="WEB-INF/logging.properties"/>
     </system-properties>
+
 </appengine-web-app>
